added one more depth for search on extensions so we can have /assets/extensions/myextension/myextension.php
The goal of this is also that some still use SVN, which needs folder level
repository control. Also, it makes shipping /assets/extensions/aiki along
side other extensions, more paletteable. It makes extensions easier to manage
from other repositories too.

The question now is how can we load extensions in the config file or other
places so we don't have to always load just in time in the db. :)