Added and fixed comments to the core, moved timezone variable to config, and made sure to check that it set in the code for future versions. Noticed that there is way too much configuration setup done all around the codebase.
there is a $aiki->get_config() that harmonizes config options, but then there
is an $aiki->processVars() which processes configuration options and aiki
markup, which is bad to have that in the aiki class. 

Think we need to have some type of configuration processor somewhere, because
there is logic for handling timezone in processVars, and then config 
processing all around the codebase. Try to do a:

grep -Rn processVars *

You will see what I mean!